Move all rAF-related code to rAF folder
Nothing changes to the external world here.
